Foundations of GfxResource binary format.

+++ LibUberRender/src/urender/common/io/base/iface/Seekable.java
@@ -3,6 +3,6 @@ package urender.common.io.base.iface;
 import java.io.IOException;
 
 public interface Seekable extends Positioned {
-    public void seek(int position) throws IOException;
-	public void setLength(int length) throws IOException;
+    public void seek(long position) throws IOException;
+	public void setLength(long length) throws IOException;
 }
